refactor: HttpBody 를 request 에서만 사용하는 객체로 변경
