+++ tomcat/src/main/java/org/apache/coyote/http11/request/HttpRequest.java
@@ -2,7 +2,6 @@ package org.apache.coyote.http11.request;
 
 import java.util.Map;
 
-import org.apache.coyote.http11.common.HttpBody;
 import org.apache.coyote.http11.common.HttpHeaders;
 import org.apache.coyote.http11.request.header.Method;
 import org.apache.coyote.http11.request.header.RequestLine;
@@ -11,15 +10,15 @@ public class HttpRequest {
 
     private final RequestLine requestLine;
     private final HttpHeaders headers;
-    private final HttpBody body;
+    private final HttpRequestBody body;
 
-    private HttpRequest(final RequestLine requestLine, final HttpHeaders headers, final HttpBody body) {
+    private HttpRequest(final RequestLine requestLine, final HttpHeaders headers, final HttpRequestBody body) {
         this.requestLine = requestLine;
         this.headers = headers;
         this.body = body;
     }
 
-    public static HttpRequest of(final RequestLine requestLine, final HttpHeaders headers, final HttpBody body) {
+    public static HttpRequest of(final RequestLine requestLine, final HttpHeaders headers, final HttpRequestBody body) {
         return new HttpRequest(requestLine, headers, body);
     }
